+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## [0.0.1] - 2026-01-06
 
 - First version
+
+## [0.0.2] - 2025-01-07
+
+- Added `Rational.parse(String value)`
+
+## [0.0.3] - 2025-01-07
+
+- Added `Decimal.stripTrailingZeros()`